core/test/com/g5/businesslogic/AccountFacadeTest.java
package com.g5.businesslogic;

import com.g5.entities.AccountEntity;
import com.g5.types.Account;
import com.g5.types.Customer;
import java.math.BigDecimal;
import java.util.List;
import javax.ejb.embeddable.EJBContainer;
import javax.naming.Context;
import javax.naming.NamingException;
import javax.persistence.EntityManager;
import javax.persistence.EntityManagerFactory;
import javax.persistence.Persistence;
import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;
import static org.junit.Assert.*;

public class AccountFacadeTest {

    private final static String ACCOUNT_FACADE_LOCAL =
            "java:global/classes/AccountFacade!com.g5.businesslogic.AccountFacadeLocal";
    private final static String PERSISTENCE_UNIT = "g5-non-jta";

    private static EJBContainer container;
    private static Context context;
    private EntityManagerFactory entityManagerFactory;
    private EntityManager entityManager;
    private TestData testData;

    @BeforeClass
    public static void setUpClass() {
        container = EJBContainer.createEJBContainer();
        context = container.getContext();
    }

    @AfterClass
    public static void tearDownClass() {
        container.close();
    }

    @Before
    public void setUp() {
        entityManagerFactory = Persistence.createEntityManagerFactory(
                PERSISTENCE_UNIT);
        entityManager = entityManagerFactory.createEntityManager();
        testData = TestData.reset(entityManager);
    }

    @After
    public void tearDown() {
        entityManager.clear();
        entityManager.close();
        entityManagerFactory.close();
    }

    private AccountFacadeLocal getAccountFacade() throws NamingException {
        return (AccountFacadeLocal) context.lookup(ACCOUNT_FACADE_LOCAL);
    }

    @Test
    public void testCreateAccount() throws Exception {
        System.out.println("createAccount");

        Customer customer = testData.getCustomers().get(0);

        AccountFacadeLocal facade = getAccountFacade();

        long minTime = System.currentTimeMillis() / 1000 * 1000;

        Account result = facade.createAccount(customer.getId());

        long maxTime = (System.currentTimeMillis() / 1000 + 1) * 1000;

        entityManager.refresh(customer);

        List<AccountEntity> accounts = entityManager.createNamedQuery(
                "Account.findByCustomer", AccountEntity.class).setParameter(
                        "customer", customer).getResultList();

        assertEquals(testData.getAccounts().size() + 1, accounts.size());

        Account account = accounts.get(accounts.size() - 1);

        assertEquals(account.getId(), result.getId());

        assertTrue(result.getCreationDate().getTime() >= minTime);
        assertTrue(result.getCreationDate().getTime() <= maxTime);
        assertEquals(BigDecimal.ZERO, result.getBalance());
    }

    @Test
    public void testFindAccountById() throws Exception {
        System.out.println("findAccountById");

        Account account = testData.getAccounts().get(0);

        entityManager.refresh(account);

        AccountFacadeLocal facade = getAccountFacade();

        Account result = facade.findAccountById(account.getId());

        assertEquals(account.getId(), result.getId());
        // TODO: This assertion fails. result date is slightly older
        //assertEquals(account.getCreationDate(), result.getCreationDate());
        assertEquals(account.getBalance(), result.getBalance());
    }

    @Test
    public void testCloseAndReopenAccount() throws Exception {
        System.out.println("closeAndReopenAccount");

        Account account = testData.getAccounts().get(0);

        AccountFacadeLocal facade = getAccountFacade();

        facade.closeAccount(account.getId());

        entityManager.refresh(account);

        assertFalse(account.isOpen());

        facade.reopenAccount(account.getId());

        entityManager.refresh(account);

        assertTrue(account.isOpen());
    }

}
